requirements.txt
langchain-openai 
pandas 
numpy 
langchain
python-dotenv
langchain_community 
langchain_core
streamlit
langchain_groq